perf(middleware): prioritize finding the authorized role
The mechanism in this commit checks a permission combination on
different roles rather than checking a role if it matches a certain
combination.

Roles have been assumed to match different combinations and may not work
on certain policy. On the other hand, combination on each policy are
closely related and only needs one role to match.